requirements.txt
flask
python-dotenv
requests
textblob
translatepy
scikit-learn
pandas
nltk
